Исправлены файлы для тестов
