+++ utils.py
@@ -17,15 +17,6 @@ def load_image_from_file(file):
     return img
 
 
-def image_classification(picture, processor, model):
-    """Обработка и распознавание изображения."""
-    inputs = processor(images=picture, return_tensors="pt")
-    outputs = model(**inputs)
-    logits = outputs.logits
-    predicted_class_idx = logits.argmax(-1).item()
-    return model.config.id2label[predicted_class_idx]
-
-
 def translate_text(text, from_language, to_language):
     """Перевод текста с одного языка на другой."""
     return ts.translate_text(text, translator="bing", from_language=from_language, to_language=to_language)
